+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/extensions/CalculatedDateTime.kt
@@ -42,13 +42,14 @@ fun Iterable<CalculatedDateTime>.asPairs(): List<CalculatedPeriod> {
     return result
 }
 
-val CalculatedDateTime.nearInPast: DateTime
-    get() {
-        val asNow = asNow
-
-        return if (asNow.isBeforeNow) {
-            asNow
-        } else {
-            asNow.minus(futureDifference)
-        }
+fun CalculatedPeriod.isBetween(dateTime: DateTime): Boolean {
+    var firstDateTime = first.asFor(dateTime)
+    val secondDateTime = second.asFutureFor(dateTime)
+    if (firstDateTime.isAfter(secondDateTime)) {
+        firstDateTime = first.asPastFor(dateTime)
     }
+    return dateTime.isAfter(firstDateTime) && dateTime.isBefore(secondDateTime)
+}
+
+val CalculatedPeriod.nowIsBetween: Boolean
+    get() = isBetween(DateTime.now())